@@ -25,30 +25,31 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@types/chalk": "^0.4.31",
-    "@types/deep-equal": "^1.0.0",
-    "@types/minimist": "^1.1.29",
-    "@types/mz": "0.0.30",
-    "@types/request": "0.0.34",
-    "@types/request-promise": "^4.1.33",
-    "async": "^2.1.4",
-    "chalk": "^2.1.0",
+    "@types/bluebird": "^3.5.18",
+    "@types/chalk": "^2.2.0",
+    "@types/deep-equal": "^1.0.1",
+    "@types/minimist": "1.2.0",
+    "@types/mz": "0.0.32",
+    "@types/request": "2.0.7",
+    "@types/request-promise": "^4.1.39",
+    "async": "^2.6.0",
+    "chalk": "^2.3.0",
     "deep-equal": "^1.0.1",
+    "flat-map-polyfill": "^0.3.8",
     "minimist": "^1.2.0",
-    "mz": "^2.6.0",
-    "request": "^2.79.0",
-    "request-promise": "^4.1.1",
-    "flat-map-polyfill": "^0.3.6"
+    "mz": "^2.7.0",
+    "request": "^2.83.0",
+    "request-promise": "^4.2.2"
   },
   "devDependencies": {
-    "@types/tape": "^4.2.30",
-    "jspm": "beta",
+    "@types/tape": "^4.2.31",
+    "jspm": "^0.17.0-beta.47",
     "rimraf": "2.6.2",
     "source-map-support": "^0.5.0",
     "tape": "^4.8.0",
     "ts-node": "^3.3.0",
-    "tslint": "^5.6.0",
-    "typescript": "^2.6.0-dev.20171005"
+    "tslint": "^5.8.0",
+    "typescript": "^2.7.0-dev.20171109"
   },
   "keywords": [
     "aurelia",
